Rename GitHub-tmp to GitHub
